[MOV] im_livechat: move DocumentViewer outside of public_livechat.js
Task-2212347
